Update git clone in build-plugin.sh to use https
While using build-plugin.sh I got an error message stating that I
did not have permission to clone the repository. Users are not
required to have a github account or have special permissions to
clone the repo when using https, so I have updated the script to
use https instead of ssh to make the script more compatible and
accessible.